backend/scraper/requirements.txt
feedparser
supabase
requests
beautifulsoup4
python-dotenv